combination

+++ cpl/src/combination.rs
@@ -5,8 +5,8 @@ pub mod combination {
         fact_inv: Vec<T>,
     }
 
-    impl PrimeModCombination<u64> {
-        pub fn new(n: usize, p: u64) -> Self {
+    impl PrimeModCombination<usize> {
+        pub fn new(n: usize, p: usize) -> Self {
             // i! mod p
             let mut fact = vec![0; n + 1];
             // (i!)^{-1} mod p
@@ -19,7 +19,7 @@ pub mod combination {
             fact_inv[1] = 1;
             iinv[1] = 1;
             for i in 2..n + 1 {
-                let j = i as u64;
+                let j = i;
                 fact[i] = fact[i - 1] * j % p;
                 iinv[i] = p - iinv[(p % j) as usize] * (p / j) % p;
                 fact_inv[i] = fact_inv[i - 1] * iinv[i] % p;
@@ -27,7 +27,7 @@ pub mod combination {
             Self { p, fact, fact_inv }
         }
 
-        pub fn combination(&self, n: usize, k: usize) -> u64 {
+        pub fn combination(&self, n: usize, k: usize) -> usize {
             self.fact[n] * (self.fact_inv[k] * self.fact_inv[n - k] % self.p) % self.p
         }
     }
